+++ src/cli/program/register.onboard.ts
@@ -6,7 +6,6 @@ import { theme } from "../../../packages/terminal-core/src/theme.js";
 import { formatAuthChoiceChoicesForCli } from "../../commands/auth-choice-options.js";
 import type { GatewayDaemonRuntime } from "../../commands/daemon-runtime.js";
 import { CORE_ONBOARD_AUTH_FLAGS } from "../../commands/onboard-core-auth-flags.js";
-import { rejectOnboardingOption } from "../../commands/onboard-options.js";
 import type {
   AuthChoice,
   GatewayAuthChoice,
@@ -45,11 +44,11 @@ const MODERN_ONBOARD_OPTION_KEYS = new Set([
   "json",
 ]);
 
-function validateRecommendationParentOptions(
+async function validateRecommendationParentOptions(
   command: Command,
   runtime: RuntimeEnv,
   json = false,
-): boolean {
+): Promise<boolean> {
   const unsupported = listExplicitOptionFlagsExcept(
     command,
     json ? RECOMMENDATION_READ_PARENT_OPTIONS : NO_RECOMMENDATION_PARENT_OPTIONS,
@@ -57,6 +56,7 @@ function validateRecommendationParentOptions(
   if (unsupported.length === 0) {
     return true;
   }
+  const { rejectOnboardingOption } = await import("../../commands/onboard-options.js");
   return rejectOnboardingOption(
     { json },
     runtime,
@@ -227,12 +227,13 @@ function pickOnboardAuthOptionValues(opts: Record<string, unknown>): Partial<Onb
   };
 }
 
-export function resolveOnboardCommandOptions(
+export async function resolveOnboardCommandOptions(
   opts: Record<string, unknown>,
   command: Command,
   runtime: RuntimeEnv,
-): OnboardOptions | false {
+): Promise<OnboardOptions | false> {
   if (opts.customImageInput === true && opts.customTextInput === true) {
+    const { rejectOnboardingOption } = await import("../../commands/onboard-options.js");
     const message = "Use either --custom-image-input or --custom-text-input, not both.";
     return rejectOnboardingOption({ json: opts.json === true }, runtime, message);
   }
@@ -321,7 +322,7 @@ export function registerOnboardCommand(program: Command): void {
       const { defaultRuntime } = await import("../../runtime.js");
       await runCommandWithRuntime(defaultRuntime, async () => {
         const json = Boolean(opts.json || recommendationsCommand.parent?.opts().json);
-        if (!validateRecommendationParentOptions(command, defaultRuntime, json)) {
+        if (!(await validateRecommendationParentOptions(command, defaultRuntime, json))) {
           return;
         }
         const { onboardRecommendationsCommand } =
@@ -338,8 +339,8 @@ export function registerOnboardCommand(program: Command): void {
       const { defaultRuntime } = await import("../../runtime.js");
       await runCommandWithRuntime(defaultRuntime, async () => {
         if (
-          !validateRecommendationParentOptions(command, defaultRuntime) ||
-          !validateRecommendationParentOptions(recommendations, defaultRuntime)
+          !(await validateRecommendationParentOptions(command, defaultRuntime)) ||
+          !(await validateRecommendationParentOptions(recommendations, defaultRuntime))
         ) {
           return;
         }
@@ -356,8 +357,8 @@ export function registerOnboardCommand(program: Command): void {
       const { defaultRuntime } = await import("../../runtime.js");
       await runCommandWithRuntime(defaultRuntime, async () => {
         if (
-          !validateRecommendationParentOptions(command, defaultRuntime) ||
-          !validateRecommendationParentOptions(recommendations, defaultRuntime)
+          !(await validateRecommendationParentOptions(command, defaultRuntime)) ||
+          !(await validateRecommendationParentOptions(recommendations, defaultRuntime))
         ) {
           return;
         }
@@ -370,6 +371,7 @@ export function registerOnboardCommand(program: Command): void {
   command.action(async (opts, commandRuntime: Command) => {
     const { defaultRuntime } = await import("../../runtime.js");
     await runCommandWithRuntime(defaultRuntime, async () => {
+      const { rejectOnboardingOption } = await import("../../commands/onboard-options.js");
       const rejectOption = (message: string) =>
         rejectOnboardingOption({ json: Boolean(opts.json) }, defaultRuntime, message);
       if (opts.modern) {
@@ -416,7 +418,7 @@ export function registerOnboardCommand(program: Command): void {
         );
         return;
       }
-      const onboardingOptions = resolveOnboardCommandOptions(
+      const onboardingOptions = await resolveOnboardCommandOptions(
         opts as Record<string, unknown>,
         commandRuntime,
         defaultRuntime,
